pyproject.toml
[project]
name = "code-auditor-agent"
version = "3.2.17"
description = "Six-phase PR review pipeline and 10-phase codebase audit pipeline for Claude Code"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "pyyaml>=6.0",
]
authors = [
    { name = "Emasoft", email = "713559+Emasoft@users.noreply.github.com" },
]

[project.optional-dependencies]
dev = [
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "types-pyyaml>=6.0",
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra"

[tool.ruff]
line-length = 120
target-version = "py312"
# Exclude CPV-upstream scripts (from claude-plugins-validation — do not lint locally)
extend-exclude = [
    "scripts/validate_*.py",
    "scripts/manage_*.py",
    "scripts/standardize_*.py",
    "scripts/generate_*.py",
    "scripts/cpv_*.py",
    "scripts/lint_files.py",
    "scripts/smart_exec.py",
    "scripts/gitignore_filter.py",
    "scripts/setup_git_hooks.py",
    "scripts/setup_plugin_pipeline.py",
    "scripts/setup_marketplace_automation.py",
    "scripts/update_marketplace_metadata.py",
    "scripts/claude-plugin-install.py",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["SIM102"]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
exclude = ["scripts_dev/", "docs_dev/", "builds_dev/"]

[tool.pyright]
pythonVersion = "3.12"

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "types-pyyaml>=6.0.12.20250915",
    "pytest>=8.0",
]
